requirements/base.txt
Django>=1.7,<1.8
Pillow==2.7.0
celery==3.1.17
cssselect==0.9.1
django-appconf==0.6
django-allauth==0.18.0
django-braces==1.8.0
django-compressor==1.2
django-extensions==1.4.8
django-localflavor==1.1
django-pagination==1.0.7
django-markdown-deux==1.0.5
django-registration==0.8
django-storages-redux==1.2.3
django-tagging==0.3.1
djangorestframework==3.0.3
extras==0.0.3
fuzzywuzzy==0.4.0
gnureadline==6.3.3
lxml==3.4.1
numpy==1.9.1
psycopg2==2.5.4
python-Levenshtein==0.12.0
python-memcached==1.48
python-mimeparse==0.1.4
python-subunit==1.0.0
pytz==2014.10
redis==2.10.3
requests==2.5.1
six==1.8.0
sorl-thumbnail==12.2
sure==1.2.9
boto==2.38.0
git+https://github.com/DemocracyClub/django-uk-political-parties.git#egg=django-uk-political-parties
